using linker and creating interface so the users implement their own scheduler
